+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Getters for real and imaginary part of `APyCFloatArray`.
 - Support for logic NOT, AND, OR, and XOR in `APyCFloat`.
 - Convenience cast functions to `APyCFloat` and `APyCFloatArray`.
+- Parameter for `seed` in `APyFloatAccumulatorContext`.
 
 ### Fixed